feat: implement repository pattern
