+++ store/moviediscover/domain/MovieModule.ts
@@ -1,6 +1,28 @@
-import { Movie } from "../data/Movie";
+import { Module, VuexModule, Mutation, Action } from 'vuex-module-decorators'
+import { Movie } from '~/store/moviediscover/data/Movie'
+import movieRepository from '../data/MovieRepositroyImpl';
 
-export interface MovieModule {
-    listMovie: Array<Movie>
-    getMovies(): void
+@Module({name: 'moviediscover/domain/MovieModule', stateFactory: true, namespaced: true })
+class MovieModule extends VuexModule {
+
+    listMovie: Array<Movie> = []
+
+    get movies() {
+        return this.listMovie;
+    }
+
+    @Mutation
+    private setMovies(listMovie: Array<Movie>) {
+        this.listMovie = listMovie
+    }
+
+    @Action
+    getMovies() {
+        movieRepository
+        .fetchMovies()
+        .then( response => { this.setMovies(response.results) })
+        .catch( error => { console.log(error) })
+    }
 }
+
+export default MovieModule
