+++ utils/store-accessor.ts
@@ -1,12 +1,12 @@
 /* eslint-disable import/no-mutable-exports */
 import { Store } from 'vuex'
 import { getModule } from 'vuex-module-decorators'
-import MovieModuleImpl from '~/store/moviediscover/domain/MovieModuleImpl'
+import MovieModule from '~/store/moviediscover/domain/MovieModule'
 
-let movieModule: MovieModuleImpl
+let movieModule: MovieModule
 
 function initialiseStores(store: Store<any>): void {
-  movieModule = getModule(MovieModuleImpl, store)
+  movieModule = getModule(MovieModule, store)
 }
 
 export { initialiseStores, movieModule }
